backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
feedparser==6.0.10
pytrends==4.9.2
lightgbm==4.3.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
aiohttp==3.9.1
google-api-python-client==2.116.0
selenium==4.17.0
webdriver-manager==4.0.1
kagglehub==0.2.5
